+++ tests/static-methods.test.ts
@@ -1,12 +1,14 @@
-import { COLORS, FUNCTIONS, ColorProps } from './data/data';
+import { COLORS, FUNCTIONS, ColorProps, ColorTranslator } from './data/data';
+
+type Props = keyof typeof COLORS[0];
 
 //Iterate over the colors
 COLORS.forEach((item: ColorProps): void => {
 
-    const colors = Object.keys(item);
+    const colors = Object.keys(item) as Props[];
 
     // Iterate over the color models
-    colors.forEach((color1: keyof ColorProps): void => {
+    colors.forEach((color1: Props): void => {
 
         describe(`ColorTranslator dynamic tests from ${color1}: ${JSON.stringify(item[color1])}`, (): void => {
 
@@ -14,7 +16,7 @@ COLORS.forEach((item: ColorProps): void => {
             const convertString = typeof convert === 'string' ? convert : JSON.stringify(convert);
 
             // Iterate again the color models
-            colors.forEach((color2: keyof ColorProps): void => {
+            colors.forEach((color2: Props): void => {
 
                 const functionName = FUNCTIONS[color2].name;
                 const functionCall = FUNCTIONS[color2].func;